requirements.txt
fastapi
uvicorn
pandas
scikit-learn
joblib
requests
python-dotenv